hibernate relations corrected transaction propagation bug resolved simulation end sentinel bug resolved first successful run of simulation

+++ src/main/java/com/cropster/roastingsimulation/greencoffee/service/GreenCoffeeService.java
@@ -14,7 +14,7 @@ public interface GreenCoffeeService {
 
     GreenCoffee getRandomFromFacility(Facility facility);
 
-    GreenCoffee reduceInStockAmount(GreenCoffee greenCoffee, int consumed);
+    GreenCoffee reduceInStockAmount(GreenCoffee greenCoffee, double consumed);
 
     GreenCoffee retrieve(String name, Facility facility);
